Add expenses

+++ database/migrations/2017_11_13_194530_create_expenses_table.php
@@ -18,7 +18,7 @@ class CreateExpensesTable extends Migration
             $table->unsignedInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->string('description');
-            $table->bigInteger('amount');
+            $table->decimal('amount');
             $table->timestamps();
         });
     }
